move most of siad to api package
