+++ api/daemon_test.go
