+++ api/daemonapi_test.go
