+++ api/miner_test.go
