+++ api/wallet_test.go
