+++ siad/daemon.go
@@ -1,12 +1,9 @@
 package main
 
 import (
-	"errors"
-	"os"
 	"path/filepath"
 
-	"github.com/stretchr/graceful"
-
+	"github.com/NebulousLabs/Sia/api"
 	"github.com/NebulousLabs/Sia/consensus"
 	"github.com/NebulousLabs/Sia/modules"
 	"github.com/NebulousLabs/Sia/modules/gateway"
@@ -28,137 +25,46 @@ type DaemonConfig struct {
 	SiaDir string
 }
 
-// The daemon is essentially a collection of modules and an API server to talk
-// to them all.
 type daemon struct {
-	state   *consensus.State
-	gateway modules.Gateway
-	host    modules.Host
-	hostdb  modules.HostDB
-	miner   modules.Miner
-	renter  modules.Renter
-	tpool   modules.TransactionPool
-	wallet  modules.Wallet
-
-	minerUpdates  <-chan struct{}
-	walletUpdates <-chan struct{}
-
-	apiServer *graceful.Server
+	srv *api.Server
 }
 
 // newDaemon will take the config struct and create a new daemon based on the
 // parameters.
 func newDaemon(config DaemonConfig) (d *daemon, err error) {
+	state = consensus.CreateGenesisState()
+	gateway, err = gateway.New(config.RPCAddr, state, filepath.Join(config.SiaDir, "gateway"))
+	if err != nil {
+		return
+	}
+	tpool, err = transactionpool.New(state, gateway)
+	if err != nil {
+		return
+	}
+	wallet, err = wallet.New(state, tpool, filepath.Join(config.SiaDir, "wallet"))
+	if err != nil {
+		return
+	}
+	miner, err = miner.New(state, gateway, tpool, wallet)
+	if err != nil {
+		return
+	}
+	host, err = host.New(state, tpool, wallet, filepath.Join(config.SiaDir, "host"))
+	if err != nil {
+		return
+	}
+	hostdb, err = hostdb.New(state, gateway)
+	if err != nil {
+		return
+	}
+	renter, err = renter.New(state, gateway, hostdb, wallet, filepath.Join(config.SiaDir, "renter"))
+	if err != nil {
+		return
+	}
+
 	d = new(daemon)
-
-	// Create a folder for each module in siaDir.
-	err = createSubdirs(config.SiaDir)
+	d.srv, err = api.NewServer(config.APIAddr, state, gateway, tpool, wallet, miner, host, hostdb, renter)
 	if err != nil {
 		return
 	}
-
-	d.state = consensus.CreateGenesisState()
-	d.gateway, err = gateway.New(config.RPCAddr, d.state, filepath.Join(config.SiaDir, "gateway"))
-	if err != nil {
-		return
-	}
-	d.tpool, err = transactionpool.New(d.state, d.gateway)
-	if err != nil {
-		return
-	}
-	d.wallet, err = wallet.New(d.state, d.tpool, filepath.Join(config.SiaDir, "wallet"))
-	if err != nil {
-		return
-	}
-	d.miner, err = miner.New(d.state, d.gateway, d.tpool, d.wallet)
-	if err != nil {
-		return
-	}
-	d.host, err = host.New(d.state, d.tpool, d.wallet, filepath.Join(config.SiaDir, "host"))
-	if err != nil {
-		return
-	}
-	d.hostdb, err = hostdb.New(d.state, d.gateway)
-	if err != nil {
-		return
-	}
-	d.renter, err = renter.New(d.state, d.gateway, d.hostdb, d.wallet, filepath.Join(config.SiaDir, "renter"))
-	if err != nil {
-		return
-	}
-
-	// Subscribe to the miner and wallet to keep all parts of the daemon
-	// synchronized.
-	d.minerUpdates = d.miner.MinerSubscribe()
-	d.walletUpdates = d.wallet.WalletSubscribe()
-
-	// Register RPCs for each module
-	d.gateway.RegisterRPC("AcceptBlock", d.acceptBlock)
-	d.gateway.RegisterRPC("AcceptTransaction", d.acceptTransaction)
-	d.gateway.RegisterRPC("HostSettings", d.host.Settings)
-	d.gateway.RegisterRPC("NegotiateContract", d.host.NegotiateContract)
-	d.gateway.RegisterRPC("RetrieveFile", d.host.RetrieveFile)
-
-	d.initAPI(config.APIAddr)
-
-	return
-}
-
-func (d *daemon) updateWait() {
-	<-d.minerUpdates
-	<-d.walletUpdates
-}
-
-func createSubdirs(rootDir string) error {
-	subdirs := []string{"gateway", "wallet", "host", "renter"}
-	for _, subdir := range subdirs {
-		err := os.MkdirAll(filepath.Join(rootDir, subdir), 0777)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// TODO: move this to the state module?
-func (d *daemon) acceptBlock(conn modules.NetConn) error {
-	var b consensus.Block
-	err := conn.ReadObject(&b, consensus.BlockSizeLimit)
-	if err != nil {
-		return err
-	}
-
-	err = d.state.AcceptBlock(b)
-	if err == consensus.ErrOrphan {
-		go d.gateway.Synchronize(conn.Addr())
-		return err
-	} else if err != nil {
-		return err
-	}
-
-	// Check if b is in the current path.
-	height, exists := d.state.HeightOfBlock(b.ID())
-	if !exists {
-		if consensus.DEBUG {
-			panic("could not get the height of a block that did not return an error when being accepted into the state")
-		}
-		return errors.New("state malfunction")
-	}
-	currentPathBlock, exists := d.state.BlockAtHeight(height)
-	if !exists || b.ID() != currentPathBlock.ID() {
-		return errors.New("block added, but it does not extend the state height")
-	}
-
-	d.gateway.RelayBlock(b)
-	return nil
-}
-
-// TODO: move this to the tpool module?
-func (d *daemon) acceptTransaction(conn modules.NetConn) error {
-	var t consensus.Transaction
-	err := conn.ReadObject(&t, consensus.BlockSizeLimit)
-	if err != nil {
-		return err
-	}
-	return d.tpool.AcceptTransaction(t)
 }
